# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "properties": {
          "approver": {
            "title": "Approver",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "verdict",
          "approver"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
      "AuditKind": {
        "enum": [
          "observation",
          "hypothesis",
          "policy_evaluation",
          "approval",
          "execution_attempt",
          "execution_result",
          "verification",
          "escalation"
        ],
        "title": "AuditKind",
        "type": "string"
      },
      "AuditRecord": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "incident_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Ref"
          },
          "kind": {
            "$ref": "#/components/schemas/AuditKind"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "actor",
          "summary"
        ],
        "title": "AuditRecord",
        "type": "object"
      },
      "CausalLink": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/EntityRef"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "transition": {
            "title": "Transition",
            "type": "string"
          }
        },
        "required": [
          "entity",
          "transition",
          "evidence"
        ],
        "title": "CausalLink",
        "type": "object"
      },
...
      "EntityRef": {
        "description": "Reference to a topology node. Events and the graph share one namespace.",
        "properties": {
          "cluster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/EntityKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace"
          }
        },
        "required": [
          "kind",
          "id",
          "name"
        ],
        "title": "EntityRef",
        "type": "object"
      },
      "Environment": {
        "enum": [
          "dev",
          "staging",
          "prod"
        ],
        "title": "Environment",
        "type": "string"
      },
      "EvaluateRequest": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "agent_risk_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Risk Limit"
          },
          "blast_radius_entities": {
            "default": 0,
            "title": "Blast Radius Entities",
            "type": "integer"
          },
          "blast_radius_users": {
            "default": 0,
            "title": "Blast Radius Users",
            "type": "integer"
          },
          "diagnostic_confidence": {
            "default": 1.0,
            "title": "Diagnostic Confidence",
            "type": "number"
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Environment"
              },
              {
                "type": "null"
              }
            ]
          },
          "executed_here_before": {
            "default": true,
            "title": "Executed Here Before",
            "type": "boolean"
          },
          "incident_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Ref"
          }
        },
        "required": [
          "action_id"
        ],
        "title": "EvaluateRequest",
        "type": "object"
      },
      "ExecuteRequest": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "actor": {
            "default": "human:operator",
            "title": "Actor",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Params",
            "type": "object"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "action_id",
          "verdict"
        ],
        "title": "ExecuteRequest",
        "type": "object"
      },
      "Execution": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "step_order": {
            "title": "Step Order",
            "type": "integer"
          },
          "succeeded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Succeeded"
          },
          "verdict_snapshot": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "step_order",
          "action_id",
          "started_at",
          "verdict_snapshot"
        ],
        "title": "Execution",
        "type": "object"
      },
      "ExecutionResult": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "finished_at": {
            "format": "date-time",
            "title": "Finished At",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "succeeded": {
            "title": "Succeeded",
            "type": "boolean"
          }
        },
        "required": [
          "action_id",
          "started_at",
          "finished_at",
          "succeeded",
          "dry_run",
          "output"
        ],
        "title": "ExecutionResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Hypothesis": {
        "description": "A causal explanation.\n\n`evidence` is required and non-empty: a hypothesis that cannot cite the\ntelemetry supporting it is suppressed, not surfaced at low confidence\n(the Grounding ADR). `contradicted_by` lets the system show its own doubt.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "contradicted_by": {
            "items": {
              "type": "string"
            },
            "title": "Contradicted By",
            "type": "array"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "mechanism": {
            "items": {
              "type": "string"
            },
            "title": "Mechanism",
            "type": "array"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "statement",
          "confidence",
          "evidence"
        ],
        "title": "Hypothesis",
        "type": "object"
      },
      "Impact": {
        "properties": {
          "affected_services": {
            "items": {
              "type": "string"
            },
            "title": "Affected Services",
            "type": "array"
          },
          "blast_radius_entities": {
            "default": 0,
            "title": "Blast Radius Entities",
            "type": "integer"
          },
          "estimated_users_affected": {
            "default": 0,
            "title": "Estimated Users Affected",
            "type": "integer"
          }
        },
        "title": "Impact",
        "type": "object"
      },
      "Incident": {
        "properties": {
          "affected_entities": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "title": "Affected Entities",
            "type": "array"
          },
          "causal_chain": {
            "items": {
              "$ref": "#/components/schemas/CausalLink"
            },
            "title": "Causal Chain",
            "type": "array"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "executions": {
            "items": {
              "$ref": "#/components/schemas/Execution"
            },
            "title": "Executions",
            "type": "array"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/Hypothesis"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/PlanStep"
            },
            "title": "Plan",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/IncidentSeverity"
          },
          "similar_incident_ids": {
            "items": {
              "type": "string"
            },
            "title": "Similar Incident Ids",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/IncidentState",
            "default": "detected"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/Transition"
            },
            "title": "Transitions",
            "type": "array"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verification"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "severity",
          "opened_at"
        ],
        "title": "Incident",
        "type": "object"
      },
      "IncidentSeverity": {
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "IncidentSeverity",
        "type": "string"
      },
      "IncidentState": {
        "enum": [
          "detected",
          "correlated",
          "diagnosed",
          "planned",
          "awaiting_approval",
          "executing",
          "verifying",
          "resolved",
          "verification_failed",
          "rolled_back",
          "escalated"
        ],
        "title": "IncidentState",
        "type": "string"
      },
      "PlanStep": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "expected_post_state": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Expected Post State",
            "type": "object"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "rollback_action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollback Action Id"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "order",
          "action_id",
          "expected_post_state"
        ],
        "title": "PlanStep",
        "type": "object"
      },
      "RiskAdjustment": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "delta"
        ],
        "title": "RiskAdjustment",
        "type": "object"
      },
      "Tier": {
        "enum": [
          "autonomous",
          "approval",
          "senior",
          "denied"
        ],
        "title": "Tier",
        "type": "string"
      },
      "Transition": {
        "description": "Append-only. Every state change names its actor and justification.",
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "from_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentState"
              },
              {
                "type": "null"
              }
            ]
          },
          "justification": {
            "title": "Justification",
            "type": "string"
          },
          "to_state": {
            "$ref": "#/components/schemas/IncidentState"
          }
        },
        "required": [
          "at",
          "from_state",
          "to_state",
          "actor",
          "justification"
        ],
        "title": "Transition",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "description": "Required argument to any execution. Carries its own reasoning.",
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "adjustments": {
            "items": {
              "$ref": "#/components/schemas/RiskAdjustment"
            },
            "title": "Adjustments",
            "type": "array"
          },
          "base_risk": {
            "title": "Base Risk",
            "type": "integer"
          },
          "constraints": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "object"
          },
          "denial_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Denial Reason"
          },
          "effective_risk": {
            "title": "Effective Risk",
            "type": "integer"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "granted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted At"
          },
          "granted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted By"
          },
          "incident_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Ref"
          },
          "required_approvers": {
            "items": {
              "type": "string"
            },
            "title": "Required Approvers",
            "type": "array"
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          }
        },
        "required": [
          "action_id",
          "incident_ref",
          "base_risk",
          "adjustments",
          "effective_risk",
          "tier",
          "required_approvers"
        ],
        "title": "Verdict",
        "type": "object"
      },
      "Verification": {
        "description": "Observational, never model-judged (the Grounding ADR).",
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/VerificationCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "window_seconds": {
            "title": "Window Seconds",
            "type": "integer"
          }
        },
        "required": [
          "window_seconds"
        ],
        "title": "Verification",
        "type": "object"
      },
      "VerificationCheck": {
        "properties": {
          "expected": {
            "title": "Expected",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          }
        },
        "required": [
          "name",
          "expected"
        ],
        "title": "VerificationCheck",
        "type": "object"
      },
      "VerifyRequest": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "incident_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Ref"
          },
          "observed": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Observed",
            "type": "object"
          },
          "window_seconds": {
            "default": 60,
            "title": "Window Seconds",
            "type": "integer"
          }
        },
        "required": [
          "action_id",
          "observed"
        ],
        "title": "VerifyRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Autonomous intelligence for complex systems. Observe \u2192 Reason \u2192 Act \u2192 Verify, under policy-bounded autonomy.",
    "title": "Pashupatastra API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/actions/execute": {
      "post": {
        "operationId": "execute_action_api_v1_actions_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Action"
      }
    },
    "/api/v1/audit": {
      "get": {
        "operationId": "audit_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_ref",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Ref"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditRecord"
                  },
                  "title": "Response Audit Api V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit"
      }
    },
...
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident"
      }
    },
    "/api/v1/policy/approve": {
      "post": {
        "operationId": "approve_api_v1_policy_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verdict"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve"
      }
    },
    "/api/v1/policy/evaluate": {
      "post": {
        "description": "Dharma. The only way to obtain authorization to execute.",
        "operationId": "evaluate_policy_api_v1_policy_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verdict"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Policy"
      }
    },
    "/api/v1/verification/observe": {
      "post": {
        "operationId": "observe_api_v1_verification_observe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Observe"
      }
    }
  }
}